test: Add test that worker status is what we expect
