+++ test/e2e/cross_os/job_attachment_bundle/template.json
